# services/api/openapi/panda-atlas-v2.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/review/cases": {
      "post": {
        "operationId": "openReviewCase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenReviewCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCaseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Review"
        ]
      },
      "get": {
        "operationId": "listReviewCases",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "triage",
                "assigned",
                "waiting",
                "decision_ready",
                "incorporation_recommended",
                "closed"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCasePageDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Review"
        ]
      }
    },
...
    "/api/v2/moderation/accounts/{accountId}": {
      "get": {
        "operationId": "getModerationAccount",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationAccountDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v2/moderation/appeals": {
      "get": {
        "operationId": "listModerationAppeals",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "under_review",
                "closed"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationAppealPageDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v2/moderation/accounts/{accountId}/sanctions": {
      "post": {
        "operationId": "applyModerationSanction",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplySanctionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationSanctionDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v2/moderation/sanctions/{sanctionId}/restore": {
      "post": {
        "operationId": "restoreModerationSanction",
        "parameters": [
          {
            "name": "sanctionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreSanctionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoredSanctionDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v2/moderation/appeals/{appealCaseId}/decision": {
      "post": {
        "operationId": "decideModerationAppeal",
        "parameters": [
          {
            "name": "appealCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideAppealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationAppealDecisionDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v2/me/moderation/appeals": {
      "post": {
        "operationId": "submitModerationAppeal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAppealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationAppealDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Moderation"
        ]
      }
    },
...
  },
  "info": {
    "title": "PandaAtlas V2 API",
    "description": "Canonical NestJS V2 HTTP contract for PandaAtlas product and administration APIs.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabaseJwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OpenReviewCaseDto": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "submissionId"
        ]
      },
      "ReviewCaseDto": {
        "type": "object",
        "properties": {
          "reviewCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "submissionId": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "number",
            "minimum": 1
          },
          "state": {
            "type": "string"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "primaryAssigneeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "reviewCaseId",
          "submissionId",
          "revisionNumber",
          "state",
          "version"
        ]
      },
      "ReviewCaseQueueItemDto": {
        "type": "object",
        "properties": {
          "reviewCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "submissionId": {
            "type": "string",
            "format": "uuid"
          },
          "revisionNumber": {
            "type": "number",
            "minimum": 1
          },
          "state": {
            "type": "string"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "primaryAssigneeId": {
            "type": "string",
            "format": "uuid"
          },
          "riskLevel": {
            "type": "string"
          },
          "targetPandaId": {
            "type": "string",
            "format": "uuid"
          },
          "contributorStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "firstResponseDueAt": {
            "type": "string",
            "format": "date-time"
          },
          "slaOverdue": {
            "type": "boolean"
          },
          "queueAgeSeconds": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "reviewCaseId",
          "submissionId",
          "revisionNumber",
          "state",
          "version",
          "riskLevel",
          "createdAt",
          "updatedAt",
          "firstResponseDueAt",
          "slaOverdue",
          "queueAgeSeconds"
        ]
      },
      "ReviewCasePageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCaseQueueItemDto"
            }
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "ModerationSubjectDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "submissionRestricted": {
            "type": "boolean"
          },
          "attachmentRestricted": {
            "type": "boolean"
          },
          "notificationRestricted": {
            "type": "boolean"
          },
          "accountSuspended": {
            "type": "boolean"
          },
          "accountClosedForAbuse": {
            "type": "boolean"
          },
          "repeatAbuseCount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "accountId",
          "version",
          "submissionRestricted",
          "attachmentRestricted",
          "notificationRestricted",
          "accountSuspended",
          "accountClosedForAbuse",
          "repeatAbuseCount"
        ]
      },
      "ModerationSanctionDto": {
        "type": "object",
        "properties": {
          "sanctionId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "warning",
              "submission_restricted",
              "attachment_restricted",
              "notification_restricted",
              "account_suspended",
              "account_closed_for_abuse"
            ]
          },
          "reasonCode": {
            "type": "string"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "sanctionId",
          "accountId",
          "kind",
          "reasonCode",
          "startsAt",
          "createdAt"
        ]
      },
      "ModerationAccountDto": {
        "type": "object",
        "properties": {
          "subject": {
            "$ref": "#/components/schemas/ModerationSubjectDto"
          },
          "sanctions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationSanctionDto"
            }
          }
        },
        "required": [
          "subject",
          "sanctions"
        ]
      },
      "ModerationAppealQueueItemDto": {
        "type": "object",
        "properties": {
          "appealCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "sanctionId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "under_review",
              "closed"
            ]
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "userStatement": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "firstResponseDueAt": {
            "type": "string",
            "format": "date-time"
          },
          "slaOverdue": {
            "type": "boolean"
          },
          "ageSeconds": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "appealCaseId",
          "accountId",
          "sanctionId",
          "state",
          "version",
          "userStatement",
          "createdAt",
          "updatedAt",
          "firstResponseDueAt",
          "slaOverdue",
          "ageSeconds"
        ]
      },
      "ModerationAppealPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModerationAppealQueueItemDto"
            }
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 100
          },
          "offset": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "ApplySanctionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "warning",
              "submission_restricted",
              "attachment_restricted",
              "notification_restricted",
              "account_suspended",
              "account_closed_for_abuse"
            ]
          },
          "reasonCode": {
            "type": "string"
          },
          "internalExplanation": {
            "type": "string"
          },
          "userVisibleExplanation": {
            "type": "string"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "idempotencyKey": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "reasonCode",
          "internalExplanation",
          "userVisibleExplanation",
          "idempotencyKey"
        ]
      },
      "RestoreSanctionDto": {
        "type": "object",
        "properties": {
          "reasonCode": {
            "type": "string"
          },
          "internalExplanation": {
            "type": "string"
          },
          "userVisibleExplanation": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        },
        "required": [
          "reasonCode",
          "internalExplanation",
          "userVisibleExplanation",
          "idempotencyKey"
        ]
      },
      "RestoredSanctionDto": {
        "type": "object",
        "properties": {
          "restored": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "restored"
        ]
      },
      "DecideAppealDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "upheld",
              "modified",
              "overturned",
              "dismissed"
            ]
          },
          "internalExplanation": {
            "type": "string"
          },
          "userVisibleExplanation": {
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "internalExplanation",
          "userVisibleExplanation"
        ]
      },
      "ModerationAppealDecisionDto": {
        "type": "object",
        "properties": {
          "decisionId": {
            "type": "string",
            "format": "uuid"
          },
          "appealCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "upheld",
              "modified",
              "overturned",
              "dismissed"
            ]
          },
          "decidedByAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "decisionId",
          "appealCaseId",
          "outcome",
          "decidedByAccountId"
        ]
      },
      "SubmitAppealDto": {
        "type": "object",
        "properties": {
          "sanctionId": {
            "type": "string",
            "format": "uuid"
          },
          "userStatement": {
            "type": "string"
          }
        },
        "required": [
          "sanctionId",
          "userStatement"
        ]
      },
      "ModerationAppealDto": {
        "type": "object",
        "properties": {
          "appealCaseId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "sanctionId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "under_review",
              "closed"
            ]
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "userStatement": {
            "type": "string"
          }
        },
        "required": [
          "appealCaseId",
          "accountId",
          "sanctionId",
          "state",
          "version",
          "userStatement"
        ]
      },
...
    }
  }
}